magen: Use gdal_translate with JPEG2000 driver (jasper-1.900.1)
This removes the dependency to the unmaintained geojasper tool.

The trick to enable the use of gdal_translate is to shift the negative
values to the unsigned range (see -scale option). geojasper worked
without this since it simply ignored the (signed) input data and assumed
unsigned data. However gdal will clip negative values to 0 if the
output type is unsigned and signed jpeg2000 compression is not supported.